// pronti-frontend-main/front/src/app/modulos/contrato/servicios/dto/contrato.update-dto.ts
export interface ContratoUpdateDto {

  // nombre?: string;

  fechaInicioCobro?: string;
  numeroDeContrato?: string;
  fechaInicio?: string;
  fechaFin?: string;
  dsctoInscripcion?: number;
  dsctoPrimeraCuota?: number;
  observacion?: string;
  fechaIniciaCobro?: string;
  estado?: string;
  plazoMesSeleccionado?: number;
  version?: number;
}