// pronti-frontend-main/front/src/app/modulos/contrato/form/form-contrato.enum.ts
export enum FormContratoEnum {
  // nombre = 'nombre',

  numeroDeContrato = 'numeroDeContrato',
  fechaInicio = 'fechaInicio',
  fechaFin = 'fechaFin',
  dsctoInscripcion = 'dsctoInscripcion',
  dsctoPrimeraCuota = 'dsctoPrimeraCuota',
  observacion = 'observacion',
  fechaIniciaCobro = 'fechaIniciaCobro',
  estado = 'estado',
  plazoMesSeleccionado = 'plazoMesSeleccionado',
  version = 'version',
  // clienteEnGrupo = 'clienteEnGrupo',

}