// pronti-frontend-main/front/src/app/enums/estado-contrato.enum.ts
export enum EstadoContratoEnum {
  Registrado = 'REG',
  EnProceso = 'EPR',
  Preadjudicado = 'PAD',
  PreadjudicadoBuscando = 'PAB',
  PreadjudicadoComprado = 'PAC',
  Ofertado = 'OFR',
  Adjudicado = 'ADJ',
  CambioDePlan = 'CPN',
  CambioMonto = 'CBM',
  CambioPlazo = 'CPZ',
  Desistimiento = 'DES',
  CesionDerechos = 'CSD',
  Refinamiento = 'REF',
  Reactivacion = 'REA',
  Unificacion = 'UNI',
  Devolucion = 'DEV',
  Liquidado = 'LQD',
  EnProcesoPenalizado = 'PLD',
  PreadjudicadoAprobado = 'APR'

}